pyproject.toml
[project]
name = "morphdiffusion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "einops>=0.8.2",
    "fastapi>=0.135.1",
    "huggingface-hub>=1.7.1",
    "pydantic>=2.12.5",
    "ruff>=0.15.6",
    "setuptools<81",
    "soundfile>=0.13.1",
    "stable-audio-tools>=0.0.19",
    "torchaudio>=2.10.0",
    "torchcodec>=0.10.0",
    "uvicorn>=0.42.0",
]
